+++ backend/Dockerfile
@@ -1,0 +1,6 @@
+FROM node:alpine
+WORKDIR /usr/src/app
+COPY package*.json .
+RUN npm ci
+COPY . .
+CMD ["npm", "run", "dev"]
