+++ backend/docker-compose.yml
@@ -1,0 +1,21 @@
+services:
+  # MongoDB Service
+  mongo_db:
+    image: mongo:latest
+    restart: always
+    volumes:
+      - mongo_db:/data/db
+  # Node API Service
+  api:
+    build: .
+    ports:
+      - 8420:8420
+    environment:
+      PORT: 8420
+      MONGODB_URI: mongodb://mongo_db:27017
+      DB_NAME: my_db
+    depends_on:
+      - mongo_db
+
+volumes:
+  mongo_db: {}
